cloud-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
redis>=5.0.0
web3>=6.10.0
python-dotenv>=1.0.0
pydantic>=2.4.2 